+++ modules/butler.js
@@ -9,11 +9,71 @@ export const SHORTNAME = "cub";
 
 export const PATH = "modules/combat-utility-belt";
 
+export const WIKIPATH = "https://github.com/death-save/combat-utility-belt/wiki"
+
 export const GADGET_NAMES = {
     enhancedConditions: "enhancedConditions",
     hideNames: "hideNames",
     rerollInitiative: "rerollInitiative",
     concentrator: "concentrator"
+}
+
+export const GADGETS = {
+    giveXP: {
+        name: "Award XP",
+        info: "Provides an end of combat prompt to distribute XP from defeated hostile combatants.",
+        wiki: `${WIKIPATH}/award-xp`
+    },
+    concentrator: {
+        name: "Concentrator",
+        info: "Manages Concentration in the dnd5e game system.",
+        wiki: `${WIKIPATH}/award-xp`
+    },
+    enhancedConditions: {
+        name: "Enhanced Conditions",
+        info: "",
+        wiki: `${WIKIPATH}/enhanced-conditions`
+    },
+    hideNames: {
+        name: "Hide NPC Names",
+        info: "",
+        wiki: `${WIKIPATH}/hide-names`
+    },
+    mightySummoner: {
+        name: "Mighty Summoner",
+        info: "",
+        wiki: `${WIKIPATH}/mighty-summoner`
+    },
+    panSelect: {
+        name: "Pan/Select",
+        info: "",
+        wiki: `${WIKIPATH}/pan-select`
+    },
+    rerollInitiative: {
+        name: "Reroll Initiative",
+        info: "",
+        wiki: `${WIKIPATH}/reroll-initiative`
+    },
+    tempCombatants: {
+        name: "Temporary Combatants",
+        info: "",
+        wiki: `${WIKIPATH}/temporary-combatants`
+    },
+    triggler: {
+        name: "Triggler",
+        info: "",
+        wiki: `${WIKIPATH}/triggler`
+    },
+    actorUtility: {
+        name: "Misc Actor",
+        info: "",
+        wiki: `${WIKIPATH}/actor-misc`
+    },
+    tokenUtility: {
+        name: "Misc Token",
+        info: "",
+        wiki: null
+    }
 }
 /**
  * Stores information about well known game systems. All other systems will resolve to "other"
@@ -84,6 +144,11 @@ export const DEFAULT_CONFIG = {
         },
         icon: "modules/combat-utility-belt/icons/concentrating.svg",
         alias: "CUB: Concentrator"
+    },
+    cubPuter: {
+        id: "cub-puter",
+        title: "CUBPuter",
+        buttonId: "cub-puter-button"
     },
     enhancedConditions: {
         iconPath: `${PATH}/icons/`,
@@ -174,6 +239,9 @@ export const DEFAULT_CONFIG = {
     tempCombatants: {
         enable: false
     },
+    actorUtility: {
+        initiativeFromSheet: false
+    },
     tokenUtility: {
         autoRollHP: false,
         effectSize: {
@@ -236,13 +304,15 @@ export const FLAGS = {
 export const SETTING_KEYS = {
     concentrator: {
         enable: "enableConcentrator",
-        icon: "modules/combat-utility-belt/icons/concentrating.svg",
         outputChat: "concentratorOutputToChat",
         autoConcentrate: "autoConcentrate",
         concentrationAttribute: "concentrationAttribute",
         notifyDouble: "notifyDoubleConcentration",
         healthAttribute: "concentratorHealthAttribute", //validate necessity
         prompt: "concentratorPromptPlayer"
+    },
+    cubPuter: {
+        config: "cubPuterConfig"
     },
     enhancedConditions: {
         enable: "enableEnhancedConditions",
@@ -294,6 +364,9 @@ export const SETTING_KEYS = {
     tempCombatants: {
         enable: "enableTempCombatants",
 
+    },
+    actorUtility: {
+        initiativeFromSheet: "initiativeFromSheet"
     },
     tokenUtility: {
         mightySummoner: "enableMightySummoner",