+++ modules/settings.js
@@ -3,8 +3,23 @@ import { EnhancedConditions } from "./enhanced-conditions/enhanced-conditions.js
 import { Sidekick } from "./sidekick.js";
 import { TokenUtility } from "./utils/token.js";
 import { Concentrator } from "./concentrator.js";
+import { CUBPuter } from "./cub-puter.js";
 
 export function registerSettings() {
+
+    /* -------------------------------------------- */
+    /*                   CUBPuter                   */
+    /* -------------------------------------------- */
+
+    Sidekick.registerMenu(BUTLER.SETTING_KEYS.cubPuter.config, {
+        name: "SETTINGS.CUBPuter.ButtonN",
+        label: BUTLER.DEFAULT_CONFIG.cubPuter.title,
+        hint: "SETTINGS.CUBPuter.ButtonH",
+        icon: "fas fa-desktop",
+        type: CUBPuter,
+        restricted: true
+    });
+
     /* -------------------------------------------- */
     /*                 Concentrator                 */
     /* -------------------------------------------- */
@@ -31,7 +46,7 @@ export function registerSettings() {
         default: BUTLER.KNOWN_GAME_SYSTEMS[game.system.id] ? BUTLER.KNOWN_GAME_SYSTEMS[game.system.id].concentrationAttribute : BUTLER.KNOWN_GAME_SYSTEMS.other.concentrationAttribute,
         scope: "world",
         type: String,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -41,7 +56,7 @@ export function registerSettings() {
         default: BUTLER.KNOWN_GAME_SYSTEMS[game.system.id] ? BUTLER.KNOWN_GAME_SYSTEMS[game.system.id].healthAttribute : BUTLER.KNOWN_GAME_SYSTEMS.other.healthAttribute,
         scope: "world",
         type: String,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -51,7 +66,7 @@ export function registerSettings() {
         default: BUTLER.DEFAULT_CONFIG.concentrator.outputChat,
         scope: "world",
         type: Boolean,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -61,7 +76,7 @@ export function registerSettings() {
         default: BUTLER.DEFAULT_CONFIG.concentrator.promptRoll,
         scope: "world",
         type: Boolean,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -71,7 +86,7 @@ export function registerSettings() {
         default: BUTLER.DEFAULT_CONFIG.concentrator.autoConcentrate,
         scope: "world",
         type: Boolean,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -82,7 +97,7 @@ export function registerSettings() {
         scope: "world",
         type: String,
         choices: BUTLER.DEFAULT_CONFIG.concentrator.notifyDouble,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -124,6 +139,7 @@ export function registerSettings() {
         default: !!BUTLER.KNOWN_GAME_SYSTEMS[game.system.id] ? BUTLER.KNOWN_GAME_SYSTEMS[game.system.id].id : BUTLER.KNOWN_GAME_SYSTEMS.other.id,
         choices: Sidekick.getSystemChoices(),
         config: false,
+        apiOnly: true,
         onChange: s => {}
     });
 
@@ -164,7 +180,7 @@ export function registerSettings() {
         hint: "SETTINGS.EnhancedConditions.OutputChatH",
         scope: "world",
         type: Boolean,
-        config: true,
+        config: false,
         default: BUTLER.DEFAULT_CONFIG.enhancedConditions.outputChat,
         onChange: s => {}
     });
@@ -174,7 +190,7 @@ export function registerSettings() {
         hint: "SETTINGS.EnhancedConditions.RemoveDefaultEffectsH",
         scope: "world",
         type: Boolean,
-        config: true,
+        config: false,
         default: BUTLER.DEFAULT_CONFIG.enhancedConditions.removeDefaultEffects,
         onChange: s => {
             EnhancedConditions._updateStatusIcons();
@@ -218,7 +234,7 @@ export function registerSettings() {
         scope: "world",
         type: String,
         default: BUTLER.DEFAULT_CONFIG.hideNames.replacementString,
-        config: true,
+        config: false,
         onChange: s => {
             const enable = Sidekick.getSetting(BUTLER.SETTING_KEYS.hideNames.enable);
 
@@ -235,7 +251,7 @@ export function registerSettings() {
         scope: "world",
         type: Boolean,
         default: BUTLER.DEFAULT_CONFIG.hideNames.hideFooter,
-        config: true,
+        config: false,
         onChange: s => {
             ui.chat.render();
         }
@@ -276,7 +292,7 @@ export function registerSettings() {
         scope: "world",
         type: String,
         choices: BUTLER.DEFAULT_CONFIG.panSelect.panGM,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -287,7 +303,7 @@ export function registerSettings() {
         scope: "world",
         type: String,
         choices: BUTLER.DEFAULT_CONFIG.panSelect.panPlayers,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -308,7 +324,7 @@ export function registerSettings() {
         scope: "world",
         type: String,
         choices: BUTLER.DEFAULT_CONFIG.panSelect.panGM, //uses same options as Pan GM
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -318,7 +334,7 @@ export function registerSettings() {
         default: BUTLER.DEFAULT_CONFIG.panSelect.selectPlayers,
         scope: "world",
         type: Boolean,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -328,7 +344,7 @@ export function registerSettings() {
         default: BUTLER.DEFAULT_CONFIG.panSelect.observerDeselect,
         scope: "world",
         type: Boolean,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -352,7 +368,7 @@ export function registerSettings() {
         scope: "world",
         type: Boolean,
         default: BUTLER.DEFAULT_CONFIG.rerollInitiative.rerollTempCombatants,
-        config: true,
+        config: false,
         onChange: s => {}
     });
 
@@ -373,6 +389,20 @@ export function registerSettings() {
     });
 
     /* -------------------------------------------- */
+    /*                 ActorUtility                 */
+    /* -------------------------------------------- */
+
+    Sidekick.registerSetting(BUTLER.SETTING_KEYS.actorUtility.initiativeFromSheet, {
+        name: "SETTINGS.ActorUtility.InitiativeFromSheetN",
+        hint: "SETTINGS.ActorUtility.InitiativeFromSheetH",
+        default: BUTLER.DEFAULT_CONFIG.actorUtility.initiativeFromSheet,
+        scope: "world",
+        type: Boolean,
+        config: false,
+        onChange: s => {}
+    })
+
+    /* -------------------------------------------- */
     /*                 TokenUtility                 */
     /* -------------------------------------------- */
 
@@ -382,7 +412,7 @@ export function registerSettings() {
         default: BUTLER.DEFAULT_CONFIG.tokenUtility.autoRollHP,
         scope: "world",
         type: Boolean,
-        config: true,
+        config: false,
         onChange: s => {}
     });